Move bit_stream, move_to_front and huffman_codec.
bit_stream, move_to_front and huffman_codec are only used by
source/tools. Move into that directory to make the usage clearer.
