+++ BUILD.gn
@@ -326,13 +326,10 @@ static_library("spvtools") {
     "source/text.h",
     "source/text_handler.cpp",
     "source/text_handler.h",
-    "source/util/bit_stream.cpp",
-    "source/util/bit_stream.h",
     "source/util/bit_vector.cpp",
     "source/util/bit_vector.h",
     "source/util/bitutils.h",
     "source/util/hex_float.h",
-    "source/util/hufman_codec.h",
     "source/util/ilist.h",
     "source/util/ilist_node.h",
     "source/util/move_to_front.h",
