+++ source/comp/markv_encoder.h
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include "source/comp/bit_stream.h"
 #include "source/comp/markv.h"
 #include "source/comp/markv_codec.h"
 #include "source/comp/markv_logger.h"
-#include "source/util/bit_stream.h"
 
 #ifndef SOURCE_COMP_MARKV_ENCODER_H_
 #define SOURCE_COMP_MARKV_ENCODER_H_
@@ -152,7 +152,7 @@ class MarkvEncoder : public MarkvCodec {
   MarkvCodecOptions options_;
 
   // Bit stream where encoded instructions are written.
-  utils::BitWriterWord64 writer_;
+  BitWriterWord64 writer_;
 
   // If not nullptr, disassembled instruction lines will be written to comments.
   // Format: \n separated instruction lines, no header.
