+++ test/CMakeLists.txt
@@ -198,18 +198,23 @@ add_spvtools_unittest(
 add_spvtools_unittest(
   TARGET bit_stream
   SRCS bit_stream.cpp
+      ${CMAKE_CURRENT_SOURCE_DIR}/../source/comp/bit_stream.cpp
+      ${CMAKE_CURRENT_SOURCE_DIR}/../source/comp/bit_stream.h
   LIBS ${SPIRV_TOOLS})
 
 add_spvtools_unittest(
   TARGET huffman_codec
   SRCS huffman_codec.cpp
+      ${CMAKE_CURRENT_SOURCE_DIR}/../source/comp/bit_stream.cpp
+      ${CMAKE_CURRENT_SOURCE_DIR}/../source/comp/bit_stream.h
+      ${CMAKE_CURRENT_SOURCE_DIR}/../source/comp/huffman_codec.h
   LIBS ${SPIRV_TOOLS})
 
 add_spvtools_unittest(
   TARGET move_to_front
   SRCS move_to_front_test.cpp
-    ../source/util/move_to_front.h
-    ../source/util/move_to_front.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/../source/comp/move_to_front.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/../source/comp/move_to_front.cpp
   LIBS ${SPIRV_TOOLS})
 
 add_subdirectory(comp)
