+++ test/huffman_codec.cpp
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Contains utils for reading, writing and debug printing bit streams.
-
 #include <algorithm>
 #include <map>
 #include <sstream>
@@ -23,11 +21,11 @@
 #include <vector>
 
 #include "gmock/gmock.h"
-#include "source/util/bit_stream.h"
-#include "source/util/huffman_codec.h"
+#include "source/comp/bit_stream.h"
+#include "source/comp/huffman_codec.h"
 
 namespace spvtools {
-namespace utils {
+namespace comp {
 namespace {
 
 const std::map<std::string, uint32_t>& GetTestSet() {
@@ -315,5 +313,5 @@ TEST(Huffman, CreateFromTextU64) {
 }
 
 }  // namespace
-}  // namespace utils
+}  // namespace comp
 }  // namespace spvtools
